+++ Rendering/Camera.h
@@ -8,14 +8,14 @@ class Camera
 {
 public:
 	Camera();
-	Camera(float fov, const std::vector<size_t>& resolution);
+	Camera(float fov, const std::vector<uint16_t>& resolution);
 
 	void move(const std::vector<float>& delta);
 	void rotate(const std::vector<float>& delta);
 
 private:
 	float _fov;
-	std::vector<size_t> _resolution;
+	std::vector<uint16_t> _resolution;
 	std::vector<float> _translation;
 	std::vector<float> _rotation;
 };
